refactor booking page and fix minor styling issues
