+++ src/routes/(app)/bookings/calendar/components/grid/TimeGridEvent.svelte
@@ -1,11 +1,15 @@
 <script lang="ts">
   import * as Avatar from "$lib/components/ui/avatar";
   import type { CalendarEventExternal } from "@schedule-x/calendar";
-  import { getTime } from "./utils";
+  import { getTime } from "../../utils";
 
   const { calendarEvent }: { calendarEvent: CalendarEventExternal } = $props();
+
+  const startTime = $derived(getTime(calendarEvent.start));
+  const endTime = $derived(getTime(calendarEvent.end));
 </script>
 
+<!-- TODO: Change colour based on category -->
 <div
   class="bg-primary/15 text-primary before:bg-primary relative size-full cursor-pointer overflow-hidden rounded-md py-1 pl-2
          transition-all duration-200 ease-out before:absolute before:top-0 before:left-0
@@ -14,12 +18,14 @@
 >
   <div class="flex min-w-0 flex-col">
     <span class="text-[0.6rem] font-medium">
-      {getTime(calendarEvent.start)} - {getTime(calendarEvent.end)}
+      {startTime} - {endTime}
     </span>
+
     <span class="mt-0.5 truncate text-base font-bold">
       {calendarEvent.location}
     </span>
   </div>
+
   <div class="mt-0.5 flex items-center justify-center gap-1">
     <Avatar.Root class="size-3">
       <Avatar.Image
@@ -29,6 +35,7 @@
       />
       <Avatar.Fallback class="text-[0.5rem]">CN</Avatar.Fallback>
     </Avatar.Root>
+
     <span class="w-full truncate rounded text-xs font-medium text-nowrap">
       {calendarEvent.people}
     </span>
