+++ src/routes/(app)/bookings/calendar/components/modal/Duration.svelte
@@ -1,0 +1,43 @@
+<script lang="ts">
+  import * as m from "$paraglide/messages";
+  import { Clock } from "@lucide/svelte";
+  import { getTime, days, months } from "../../utils";
+  import Section from "./Section.svelte";
+
+  type PropDate = Temporal.ZonedDateTime | Temporal.PlainDate;
+
+  const { startDate, endDate }: { startDate: PropDate; endDate: PropDate } =
+    $props();
+
+  const startTime = $derived(getTime(startDate));
+  const startDay = $derived(startDate.day);
+  const startWeekday = $derived(days[startDate.dayOfWeek - 1]);
+  const startMonth = $derived(months[startDate.month - 1]);
+  const startYear = $derived(startDate.year);
+
+  const endTime = $derived(getTime(endDate));
+  const endDay = $derived(endDate.day);
+  const endMonth = $derived(months[endDate.month - 1]);
+  const endYear = $derived(endDate.year);
+
+  let header = $state("");
+  let description = $state("");
+
+  $effect(() => {
+    const isSameDay = startDay === endDay;
+    const isSameYear = startYear === endYear;
+
+    if (isSameDay) {
+      header = `${startDay} ${startMonth} ${startYear}`;
+      description = `${startTime} - ${endTime}`;
+    } else {
+      header = isSameYear
+        ? `${startDay} ${startMonth} - ${endDay} ${endMonth} ${endYear}`
+        : `${startDay} ${startMonth} ${startYear} - ${endDay} ${endMonth} ${endYear}`;
+
+      description = `${m.events_startTime()} ${startWeekday} @ ${startTime}`;
+    }
+  });
+</script>
+
+<Section Icon={Clock} {header} {description} />
