+++ src/routes/(app)/bookings/calendar/utils.ts
